requirements.txt
PyYAML
termcolor
requests
